# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/anime/detail-alt": {
            "post": {
                "description": "This endpoint will scrape the detail instead of fetching from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anime"
                ],
                "summary": "Anime detail alternative",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webscraper.TitleSearchResult"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.baseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Anime"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/anime/detail/{anime_id}": {
            "get": {
                "description": "By providing anime id, this endpoint will return the details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anime"
                ],
                "summary": "Get anime detail",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Anime ID",
                        "name": "anime_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.baseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Anime"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/anime/recent": {
            "get": {
                "description": "This endpoint supports pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anime"
                ],
                "summary": "Get recently updated / released anime",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Desired page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.baseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.RecentAnime"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/anime/search": {
            "get": {
                "description": "Will find all anime with matching keywords",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anime"
                ],
                "summary": "Search anime title",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Naruto",
                        "description": "Keywords of anime title",
                        "name": "keywords",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.baseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/webscraper.TitleSearchResult"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/anime/sync": {
            "post": {
                "description": "To sync anime data from client to server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anime"
                ],
                "summary": "Sync anime",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.syncAnimeRequest"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.baseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.SyncAnime"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/anime/video-url": {
            "post": {
                "description": "Fetch video URL by episode endpoint",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anime"
                ],
                "summary": "Video URL",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.videoURLRequest"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.baseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.videoURLResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/bookmark": {
            "post": {
                "description": "User can receive notification from apps when their bookmarked anime has new update",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmark"
                ],
                "summary": "Save bookmark",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.saveBookmarkRequest"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.baseResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "User can delete bookmark to opt out from notification in the future",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmark"
                ],
                "summary": "Delete bookmark",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.deleteBookmarkRequest"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.baseResponse"
                        }
                    }
                }
            }
        },
        "/token/refresh": {
            "get": {
                "description": "Refresh token will have later expiry date than auth token so user can reissue auth token whenever expired",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "summary": "Generate auth token by providing refresh token",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your refresh token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.baseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.refreshTokenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.baseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "controller.deleteBookmarkRequest": {
            "type": "object",
            "required": [
                "anime_id",
                "user_token"
            ],
            "properties": {
                "anime_id": {
                    "type": "integer"
                },
                "user_token": {
                    "type": "string"
                }
            }
        },
        "controller.refreshTokenResponse": {
            "type": "object",
            "properties": {
                "auth_token": {
                    "type": "string"
                }
            }
        },
        "controller.saveBookmarkRequest": {
            "type": "object",
            "required": [
                "anime_id",
                "latest_episode",
                "user_token"
            ],
            "properties": {
                "anime_id": {
                    "type": "integer"
                },
                "latest_episode": {
                    "type": "string"
                },
                "user_token": {
                    "type": "string"
                }
            }
        },
        "controller.syncAnimeRequest": {
            "type": "object",
            "required": [
                "anime_ids"
            ],
            "properties": {
                "anime_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "controller.videoURLRequest": {
            "type": "object",
            "required": [
                "episode_endpoint"
            ],
            "properties": {
                "episode_endpoint": {
                    "type": "string",
                    "example": "/naruto-episode-1"
                }
            }
        },
        "controller.videoURLResponse": {
            "type": "object",
            "properties": {
                "video_url": {
                    "type": "string"
                }
            }
        },
        "domain.Anime": {
            "type": "object",
            "properties": {
                "airing_year": {
                    "type": "string"
                },
                "episodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Episode"
                    }
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "latest_episode": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Episode": {
            "type": "object",
            "properties": {
                "endpoint": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "domain.RecentAnime": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "latest_episode": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.SyncAnime": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "latest_episode": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "webscraper.TitleSearchResult": {
            "type": "object",
            "required": [
                "endpoint",
                "title"
            ],
            "properties": {
                "endpoint": {
                    "type": "string",
                    "example": "/category/naruto"
                },
                "title": {
                    "type": "string",
                    "example": "naruto"
                }
            }
        }
    }
}